requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic[email]>=2.9.0
pydantic-settings>=2.2.0
asyncpg>=0.29.0
cryptography>=42.0.0
python-multipart>=0.0.9
email-validator>=2.1.0
httpx>=0.27.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
